+++ lib/src/core/google_map.state.dart
@@ -46,7 +46,7 @@ class GoogleMapState extends GoogleMapStateBase {
     String icon,
     String info,
     String infoSnippet,
-    VoidCallback onTap,
+    ValueChanged<String> onTap,
     VoidCallback onInfoWindowTap,
   }) =>
       throw UnimplementedError();
@@ -55,6 +55,7 @@ class GoogleMapState extends GoogleMapStateBase {
   void addPolygon(
     String id,
     Iterable<GeoCoord> points, {
+    ValueChanged<String> onTap,
     Color strokeColor = const Color(0x000000),
     double strokeOpacity = 0.8,
     double strokeWidth = 1,
@@ -76,6 +77,7 @@ class GoogleMapState extends GoogleMapStateBase {
   void editPolygon(
     String id,
     Iterable<GeoCoord> points, {
+    ValueChanged<String> onTap,
     Color strokeColor = const Color(0x000000),
     double strokeOpacity = 0.8,
     double strokeWeight = 1,