+++ lib/src/core/map_operations.dart
@@ -4,6 +4,7 @@
 
 import 'dart:ui' show Color, VoidCallback;
 
+import 'package:flutter/foundation.dart' show ValueChanged;
 import 'package:google_directions_api/google_directions_api.dart'
     show GeoCoord, GeoCoordBounds;
 
@@ -86,7 +87,7 @@ abstract class MapMarkers {
     String icon,
     String info,
     String infoSnippet,
-    VoidCallback onTap,
+    ValueChanged<String> onTap,
     VoidCallback onInfoWindowTap,
   });
 
@@ -141,6 +142,7 @@ abstract class MapPolygons {
   void addPolygon(
     String id,
     Iterable<GeoCoord> points, {
+    ValueChanged<String> onTap,
     Color strokeColor = const Color(0x000000),
     double strokeOpacity = 0.8,
     double strokeWidth = 1,
@@ -156,6 +158,7 @@ abstract class MapPolygons {
   void editPolygon(
     String id,
     Iterable<GeoCoord> points, {
+    ValueChanged<String> onTap,
     Color strokeColor = const Color(0x000000),
     double strokeOpacity = 0.8,
     double strokeWeight = 1,