+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [3.7.0]
+
+* Changed onTap VoidCallback to onTap ValueChanged with marker ID as an argument
+* Added onTap for polygon with polygon ID as an argument
+
 ## [3.6.0]
 
 * Added infoSnippet as an argument to addMarker method